requirements.txt
python-dotenv==0.20.0
fastapi==0.79.0
requests==2.27.1
PyMySQL==1.0.2
kafka-python==2.0.2
redis==4.3.4
elasticsearch==8.5.0
ujson==4.3.0